refactor: LottoResult
